send POST requests to hardcoded localhost:4000 (local ambi server)
I decided to not write tests for this step in the MVP

Two options I considered were:
* Finding library to provide some sort of mock that I could send requests to in tests.
  * This seemed like overkill a bit for a CLI at this stage
* Use dependency injection and define a test Client
  * It would be really cool if reqwest provided a trait for Client that we could use to create a test client. For the moment this felt like overkill as well.
Also, since we can assume the functions being used that are provided by reqwest are themselves tested within that crate maybe it does not make sense to re-test them here. We would essentially be writing a test for the `main` function.